Implementing alert messages
after the user has logged in, we want to display a message to the user that he has logged in successfully. We will use the alert component from bootstrap to display the message. We will also use the session variable to store the message and display it on the dashboard page.
same goes for the logout, sign-up and reset password.

also adjusted the css to center the auth forms.